gave the tool new simple gui
Put the file path to your list, and the website link into the ui, use the save buttons for both and then click run to run the rest of the script. A new window should be generated every time you click run. You can generate as many new tables as you want, and compare them.